cleaned up some warnings

+++ c/ast/src/string.c
@@ -126,7 +126,7 @@ static struct kn_string_t *create_string(char *str, size_t length) {
 
 size_t kn_string_length(const struct kn_string_t *string) {
 	return string->flags & KN_STRING_FL_EMBED
-		? string->embed.length
+		? (size_t) string->embed.length
 		: string->alloc.length;
 }
